% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facthelpeR-package.R
\docType{package}
\name{facthelpeR-package}
\alias{facthelpeR}
\alias{facthelpeR-package}
\title{facthelpeR: Helper Functions for Importing and Cleaning Data}
\description{
The facthelpeR package contains helper functions for importing and cleaning data.
}
\author{
\strong{Maintainer}: Jacqueline Tay \email{jacquelinejtay@gmail.com}

Other contributors:
\itemize{
  \item Emily Galvin \email{emilymgalvin@gmail.com} [contributor]
}

}
\keyword{internal}
