% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_cols.R
\name{find_col_info}
\alias{find_col_info}
\title{Find column information in a data.frame}
\usage{
find_col_info(df, cols_check, by_col)
}
\arguments{
\item{cols_check}{A character vector of column names to check}

\item{by_col}{A bare variable to summarise}
}
\description{
For a data.frame, loop through the columns provided (cols_check)
and summarise the number of non-NA entries and the groups that
contain non-NA entries (for example if want to search trial, year, or filename),
and return the first non-NA entry as an example
}
\seealso{
Other summarize column functions: 
\code{\link{select_colsfromlist}()},
\code{\link{summarize_colnames}()},
\code{\link{summarize_ncol}()},
\code{\link{summarize_variables}()}
}
\concept{summarize column functions}
