% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files_fns_wrappers.R
\name{list.sheetnames}
\alias{list.sheetnames}
\title{A wrapper for list_sheetnames to maintain backwards compatibility}
\usage{
list.sheetnames(data_folder, files, reg_ex = NULL)
}
\arguments{
\item{data_folder}{A string pointing to the parent path}

\item{files}{A vector of filenames that will be pasted to the parent path}

\item{reg_ex}{A regular expression to select sheets using}
}
\description{
A wrapper for list_sheetnames to maintain backwards compatibility
}
\seealso{
Other function rename wrappers: 
\code{\link{read.excelheader}()},
\code{\link{read.excelsheet}()},
\code{\link{read.sheets}()}
}
\concept{function rename wrappers}
