% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_sheetnames.R
\name{list_sheetnames}
\alias{list_sheetnames}
\title{List the names of all sheets within given files
Note: this function was previously list.sheetnames()}
\usage{
list_sheetnames(data_folder, files, reg_ex = NULL)
}
\arguments{
\item{data_folder}{A string pointing to the parent path}

\item{files}{A vector of filenames that will be pasted to the parent path}

\item{reg_ex}{A regular expression to select sheets using}
}
\description{
List the names of all sheets within given files
Note: this function was previously list.sheetnames()
}
