% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_direction.R
\name{make_df_index}
\alias{make_df_index}
\title{Utility function to separate a table into list elements based on a blank row or column}
\usage{
make_df_index(v)
}
\arguments{
\item{v}{A vector}
}
\description{
Utility function to separate a table into list elements based on a blank row or column
}
\keyword{internal}
