% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map.R
\name{plot_map}
\alias{plot_map}
\title{Check field map layout for errors and duplicate range/row locations.}
\usage{
plot_map(
  df,
  id_col = "trial",
  trial_id,
  plot_id = NULL,
  range = range,
  row = row
)
}
\arguments{
\item{df}{A data frame.}

\item{id_col}{A string. Column containing a unique identifier for the trial. The default is 'trial'.}

\item{trial_id}{A string. Use "all" to map all data by trial; for single trial, trial id string
to locate in id_col, i.e. "OregonCereals_SWS_ION_2023". The string to match should be exact; the function
is not currently set up to locate partial strings.}

\item{plot_id}{A string or vector; column name(s) to use as plot identifier. Currently can take up to
two inputs, i.e. c("plot", "ibloc"). This is useful for finding missing plot numbers, ibloc, entry, variety, etc.}

\item{range}{Column containing range values; the default is 'range'}

\item{row}{Column containing row locations; the default is 'row'}
}
\description{
This function can map either a single trial or all trials. When mapping all trials,
best practice is to first filter out trials without range/row. If you do this and
still receive a warning that the output will contain list cols, you have duplicates
in range/row.
}
\seealso{
Other clean column contents functions: 
\code{\link{set_values_to_na}()}
}
\concept{clean column contents functions}
