% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files_fns.R
\name{read.sheets}
\alias{read.sheets}
\title{Read in sheets}
\usage{
read.sheets(file_name, data_folder, skip_df)
}
\arguments{
\item{file_name}{A string of a filename}

\item{data_folder}{A string of the path of the folder that contains the file}

\item{skip_df}{A data.frame containing the following columns: filename, sheets, and skip,
which contain string of filename, string of sheet name, and integer of number of rows to skip, respectively
https://stackoverflow.com/questions/60898358/how-can-i-add-the-sheet-name-to-my-data-frame-as-a-variable}
}
\description{
Read in sheets
}
