% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files_fns_wrappers.R
\name{read.excelheader}
\alias{read.excelheader}
\title{A wrapper for read_excelheader to maintain backwards compatibility}
\usage{
read.excelheader(sheet_name, path, header_start, header_end, unique_names)
}
\description{
A wrapper for read_excelheader to maintain backwards compatibility
}
