% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files_fns.R, R/read_files_fns_wrappers.R
\name{read_excelsheet}
\alias{read_excelsheet}
\alias{read.excelsheet}
\title{Read one excel sheet with read_excel
This function can be incorporated into read.sheets/ read.sheets may become deprecated}
\usage{
read_excelsheet(
  path,
  sheet,
  skip,
  na,
  col_names,
  guess_max,
  complete_cases = TRUE
)

read.excelsheet(
  path,
  sheet,
  skip,
  na,
  col_names,
  guess_max,
  complete_cases = TRUE
)
}
\arguments{
\item{path}{Path to the xls/xlsx file.}

\item{sheet}{Sheet to read. Either a string (the name of a sheet), or an
integer (the position of the sheet). Ignored if the sheet is specified via
\code{range}. If neither argument specifies the sheet, defaults to the first
sheet.}

\item{skip}{Minimum number of rows to skip before reading anything, be it
column names or data. Leading empty rows are automatically skipped, so this
is a lower bound. Ignored if \code{range} is given.}

\item{na}{Character vector of strings to interpret as missing values. By
default, readxl treats blank cells as missing data.}

\item{col_names}{\code{TRUE} to use the first row as column names, \code{FALSE} to get
default names, or a character vector giving a name for each column. If user
provides \code{col_types} as a vector, \code{col_names} can have one entry per
column, i.e. have the same length as \code{col_types}, or one entry per
unskipped column.}

\item{guess_max}{Maximum number of data rows to use for guessing column
types.}

\item{complete_cases}{A logical. The default TRUE will remove empty rows}
}
\description{
Note: this function was previously read.excelsheet()
}
\seealso{
Other readin functions: 
\code{\link{list_sheetnames}()},
\code{\link{read_excelheader}()},
\code{\link{read_multsheets}()},
\code{\link{read_sheets}()}
}
\concept{readin functions}
