% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files_fns.R, R/read_files_fns_wrappers.R
\name{read_header}
\alias{read_header}
\alias{read.excelheader}
\title{Read headers that span over multiple rows}
\usage{
read_header(sheet_name, path, header_start, header_end, unique_names)

read.excelheader(sheet_name, path, header_start, header_end, unique_names)
}
\arguments{
\item{sheet_name}{A string denoting the sheetname}

\item{path}{A path to the file that contains the sheet}

\item{header_start}{A numeric denoting the row number of the start of the header
Note!: If there are any blank rows at the top of the sheet, start the row counting
at the first filled row
https://github.com/tidyverse/readxl/issues/194#issuecomment-266829259}

\item{header_end}{A numeric denoting the row number of the end of the header}

\item{unique_names}{A logical denoting whether make.unique() for the column names}
}
\description{
Note: this function was previously read.excelheader()/read_excelheader()
}
\seealso{
Other readin functions: 
\code{\link{list_sheetnames}()},
\code{\link{read_excelsheet}()},
\code{\link{read_multsheets}()},
\code{\link{read_ods_multsheets}()},
\code{\link{read_sheets}()}
}
\concept{readin functions}
