% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_multsheets.R
\name{read_multsheets}
\alias{read_multsheets}
\title{Read multiple sheets listed in a data.frame
created by list.sheetnames}
\usage{
read_multsheets(
  data_folder,
  df,
  na = c("NA"),
  col_names,
  guess_max = 1000,
  complete_cases = TRUE
)
}
\arguments{
\item{data_folder}{A string denoting the folder that contains the
files to be read in}

\item{df}{A data.frame created by the function list.sheetnames
The df can be modified to include a column "header_end," which denotes
the number of the row corresponding to the last row of the header.
header_end = 1 corresponds to the header being in only the first row and
corresponds to skipping 0 rows.
If header_end is not provided, the default
is set to 1, which will skip 0 rows.  Similarly, the df can also include a column of "list_names,"
which denote the name of the list object.  If list_names is not provided,
the default is the concatenation of the file and sheet name
Note: If header_end is NA, the file/sheet will be removed (not read in)}

\item{na}{Character vector of strings to interpret as missing values. By
default, readxl treats blank cells as missing data.}

\item{col_names}{\code{TRUE} to use the first row as column names, \code{FALSE} to get
default names, or a character vector giving a name for each column. If user
provides \code{col_types} as a vector, \code{col_names} can have one entry per
column, i.e. have the same length as \code{col_types}, or one entry per
unskipped column.}

\item{guess_max}{Maximum number of data rows to use for guessing column
types.}

\item{complete_cases}{A logical. The default TRUE will remove empty rows}
}
\description{
Read multiple sheets listed in a data.frame
created by list.sheetnames
}
\seealso{
Other readin functions: 
\code{\link{list_sheetnames}()},
\code{\link{read_excelheader}()},
\code{\link{read_excelsheet}()},
\code{\link{read_sheets}()}
}
\concept{readin functions}
