% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ods_multsheets.R
\name{read_ods_multsheets}
\alias{read_ods_multsheets}
\title{A wrapper around read_ods to read multiple sheets in a .ods file}
\usage{
read_ods_multsheets(path)
}
\arguments{
\item{path}{Path to the .ods file to import.}
}
\value{
A list of data frames--one data frame for each sheet.
}
\description{
Read in multiple sheets in a .ods file.
}
\details{
Currently, read_ods will only read one sheet at a time. read_ods_multsheets
can handle multiple sheets in one .ods workbook, or, when used with
list.files and lapply, multiple workbooks with multiple sheets.

The result is a named list of data frames, with the sheet name appended
to the basename of the filename with the file extension removed.
}
\examples{
\dontrun{
write_ods(iris[iris$Species=="setosa",], "iris.ods", sheet = "setosa")

write_ods(iris[iris$Species=="versicolor",], "iris.ods",
          sheet = "versicolor", append = TRUE)

write_ods(iris[iris$Species=="virginica",],
          "iris.ods", sheet = "virginica", append = TRUE)

read_ods_multsheets("iris.ods")

# Read multiple files with multiple sheets
files <- list.files(data_folder, recursive = TRUE, pattern = "*.ods")

# Use lapply() with the result of list.files
lapply(files, read_ods_multsheets)
}
}
\seealso{
Other readin functions: 
\code{\link{list_sheetnames}()},
\code{\link{read_excelsheet}()},
\code{\link{read_header}()},
\code{\link{read_multsheets}()},
\code{\link{read_sheets}()}
}
\concept{readin functions}
