% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files_fns.R, R/read_files_fns_wrappers.R
\name{read_sheets}
\alias{read_sheets}
\alias{read.sheets}
\title{Read in sheets}
\usage{
read_sheets(file_name, data_folder, skip_df)

read.sheets(file_name, data_folder, skip_df)
}
\arguments{
\item{file_name}{A string of a filename}

\item{data_folder}{A string of the path of the folder that contains the file}

\item{skip_df}{A data.frame containing the following columns: filename, sheets, and skip,
which contain string of filename, string of sheet name, and integer of number of rows to skip, respectively
https://stackoverflow.com/questions/60898358/how-can-i-add-the-sheet-name-to-my-data-frame-as-a-variable}
}
\description{
This function may be deprecated (read_multsheets() has many improvements
over read_sheets)
Note: this function was previously read.sheets()
}
\seealso{
Other readin functions: 
\code{\link{list_sheetnames}()},
\code{\link{read_excelheader}()},
\code{\link{read_excelsheet}()},
\code{\link{read_multsheets}()}
}
\concept{readin functions}
