% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_col.R
\name{rename.intdouble}
\alias{rename.intdouble}
\title{Rename columns for a list of data.frames based on numeric type (integer versus double)
Related to select.colsfromlist
Note: this function does not rename based on column type (int versus num).  This function
renames based on whether the numeric is a decimal or round number}
\usage{
\method{rename}{intdouble}(data_list, reg_ex, rename_int, rename_double)
}
\arguments{
\item{data_list}{A list with each element containing a data.frame}

\item{reg_ex}{A string containing a regular expression that will match the columns to be returned}

\item{rename_int}{A string to rename the column if the columns values are doubles}
}
\description{
Rename columns for a list of data.frames based on numeric type (integer versus double)
Related to select.colsfromlist
Note: this function does not rename based on column type (int versus num).  This function
renames based on whether the numeric is a decimal or round number
}
