% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_col.R
\name{rename_col}
\alias{rename_col}
\alias{rename.col}
\title{Rename columns with a data.frame containing old and new column names}
\usage{
rename_col(
  ls,
  rename_df,
  rename_col,
  old_col,
  element_col = NULL,
  rm_col = FALSE
)

rename.col
}
\arguments{
\item{ls}{A list of data.frames whose columns will be renamed}

\item{rename_df}{A data.frame that has a column containing
old column names and new column names}

\item{rename_col}{A variable name that corresponds to the column
containing new column names}

\item{old_col}{A variable name that corresponds to the column to be renamed}

\item{element_col}{A variable name that corresponds to the name of the list
element to be renamed.
The default value is NULL, meaning that the rename will be applied to
all elements in ls}

\item{rm_col}{A logical specifying whether to remove columns that have been
renamed to "rm_col" in rename_df}
}
\description{
This wrapper required using enexpr() because of the bare variables
being passed to rename_col.  See tests/understand_wrappers.R for more
details
}
\seealso{
Other clean columns functions: 
\code{\link{rm_dup_col}()}
}
\concept{clean columns functions}
