% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_dup_col.R
\name{rm_dup_col}
\alias{rm_dup_col}
\alias{rm.dup_col}
\title{Remove duplicate columns}
\usage{
rm_dup_col(df, concat_names = TRUE)

rm.dup_col(df, concat_names = TRUE)
}
\arguments{
\item{df}{A data.frame}

\item{concat_names}{A logical denoting how to rename the columns
The default is TRUE, which renames the column by concatenating the names of
all the duplicate columns
FALSE will retain the first}
}
\description{
Remove duplicate columns.
If the columns are duplicates, except one column contains NAs, then
the column with NAs is removed
#' test_dup <- duplicated(as.list(df)) This method detects duplicates
but does not take into account NAs
}
\seealso{
Other clean columns functions: 
\code{\link{rename_col}()}
}
\concept{clean columns functions}
