% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_cols.R
\name{select_colsfromlist}
\alias{select_colsfromlist}
\alias{select.colsfromlist}
\title{Select columns from data.frames stored within a list}
\usage{
select_colsfromlist(data_list, reg_ex, multnames_only = FALSE)

\method{select}{colsfromlist}(data_list, reg_ex, multnames_only = FALSE)
}
\arguments{
\item{data_list}{A list with each element containing a data.frame}

\item{reg_ex}{A string containing a regular expression that will match the columns to be returned}

\item{multnames_only}{A logical expressing whether to only return data
with more than one column corresponding to a given regular expression}
}
\description{
Select columns from data.frames stored within a list

A wrapper for select_colsfromlist to maintain backwards compatibility
}
\seealso{
Other summarize column functions: 
\code{\link{find_col_info}()},
\code{\link{summarize_colnames}()},
\code{\link{summarize_ncol}()},
\code{\link{summarize_variables}()}
}
\concept{summarize column functions}
