% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_direction.R
\name{split_df}
\alias{split_df}
\title{Split a large table into smaller tables if there are blank columns or rows}
\usage{
split_df(df, showWarning = TRUE, complexity = 1)
}
\arguments{
\item{df}{A data.frame to be split}

\item{showWarning}{A logical denoting whether to show warning message}

\item{complexity}{An integer. Starting at 1, increase complexity if tables
don't split correctly.}
}
\description{
Split a large table into smaller tables if there are blank columns or rows.
This function can detect several tables even if they are not neatly placed
(e.g. not placed in a grid)
}
\details{
If you still see entire rows or columns missing. Please increase complexity
}
\seealso{
Other split tables functions: 
\code{\link{split_direction}()}
}
\concept{split tables functions}
