% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_direction.R
\name{split_direction}
\alias{split_direction}
\title{Split a large table in one direction if there are blank columns or rows}
\usage{
split_direction(df, direction)
}
\arguments{
\item{df}{A data.frame to be split}

\item{direction}{A string denoting whether to split on a
row or column ("col" or "row")}
}
\description{
These functions were first written by @yusuzech
See stackoverflow: https://stackoverflow.com/questions/58251748/how-to-efficiently-import-multiple-excel-tables-located-in-one-sheet-into-an-r-l/58252193#58252193
and GitHub: https://github.com/yusuzech/tidyverse_notes/blob/master/utility/read_excel_tables.md
}
\seealso{
Other split tables functions: 
\code{\link{split_df}()}
}
\concept{split tables functions}
