% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_cols.R
\name{summarize_colnames}
\alias{summarize_colnames}
\alias{summarize.colnames}
\title{Create a list of column names that match a given regex for all data.frames in given list
The function returns the proportion of the data.frames that contain the column name}
\usage{
summarize_colnames(ls, col_vec)

summarize.colnames
}
\arguments{
\item{ls}{A list of data.frames}

\item{col_vec}{A vector containing string regex.
Each string will correspond to the columns to pull from each dataset
in order to summarize the number of columns associated with that regex string}
}
\description{
Create a list of column names that match a given regex for all data.frames in given list
The function returns the proportion of the data.frames that contain the column name

A wrapper for summarize_colnames to maintain backwards compatibility
}
\seealso{
Other summarize column functions: 
\code{\link{find_col_info}()},
\code{\link{select_colsfromlist}()},
\code{\link{summarize_ncol}()},
\code{\link{summarize_variables}()}
}
\concept{summarize column functions}
