% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_cols.R
\name{summarize_ncol}
\alias{summarize_ncol}
\alias{summarize.ncol}
\title{Summarise the number of columns per data.frame that match a given regex
over a list of data.frames
This function detects columns with similar names within a dataset}
\usage{
summarize_ncol(ls, col_vec)

summarize.ncol
}
\arguments{
\item{ls}{A list of data.frames}

\item{col_vec}{A vector containing string regex.
Each string will correspond to the columns to pull from each dataset
in order to summarize the number of columns associated with that regex string}
}
\description{
Summarise the number of columns per data.frame that match a given regex
over a list of data.frames
This function detects columns with similar names within a dataset

A wrapper for summarize_ncol to maintain backwards compatibility
}
\seealso{
Other summarize column functions: 
\code{\link{find_col_info}()},
\code{\link{select_colsfromlist}()},
\code{\link{summarize_colnames}()},
\code{\link{summarize_variables}()}
}
\concept{summarize column functions}
