% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_cols.R
\name{summarize_variables}
\alias{summarize_variables}
\alias{summarize.variables}
\title{Summarize information on variables that match a given regex}
\usage{
summarize_variables(data_list, reg_ex, reg_ex_exclude = NULL)

summarize.variables
}
\arguments{
\item{data_list}{A list of data.frames}

\item{reg_ex}{A regular expression to select column names
Use NULL to return all column names
It is possible to edit this function to use the above function (select.colsfromlist)}
}
\description{
Summarize information on variables that match a given regex.
This function loops through a list of data.frames and pulls out the columns
that match the regex and returns a joined data.frame of the name in the list,
the variable, variable type, and the first instance of the variable as an example.
This function is to help look at raw data in order to check which columns across files
contain the same information.
}
\seealso{
Other summarize column functions: 
\code{\link{find_col_info}()},
\code{\link{select_colsfromlist}()},
\code{\link{summarize_colnames}()},
\code{\link{summarize_ncol}()}
}
\concept{summarize column functions}
