% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_dup_col.R
\name{test.col_id_elements}
\alias{test.col_id_elements}
\title{Test whether a column has identical elements}
\usage{
test.col_id_elements(df, colname)
}
\arguments{
\item{df}{A data frame}

\item{colname}{A string. Name of column to test}
}
\description{
Test whether a column has identical elements.
The function will return TRUE if:
all values are equal,
all values are equal and there are NAs,
all values are NA.
}
\details{
https://stackoverflow.com/questions/31968623/how-to-check-whether-a-column-contains-only-identical-elements-in-r#:~:text=You\%20can\%20use\%20the\%20duplicated,column\%20contains\%20all\%20identical\%20values.&text=If\%20x\%20contains\%20NAs\%20this,mixed\%20columns\%20will\%20return\%20FALSE\%20.
}
\keyword{internal}
