% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_direction.R
\name{vec_rle}
\alias{vec_rle}
\title{Utility function to get rle as a named vector}
\usage{
vec_rle(v)
}
\arguments{
\item{v}{A vector}
}
\description{
Utility function to get rle as a named vector
}
\keyword{internal}
