% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ods_multsheets.R
\name{read_ods_multsheets}
\alias{read_ods_multsheets}
\title{A wrapper around read_ods to read multiple sheets in a .ods file}
\usage{
read_ods_multsheets(filename)
}
\arguments{
\item{filename}{File to import}
}
\value{
A list of data frames--one data frame for each sheet
}
\description{
A wrapper around read_ods to read multiple sheets in a .ods file
}
\seealso{
Other readin functions: 
\code{\link{list_sheetnames}()},
\code{\link{read_excelsheet}()},
\code{\link{read_header}()},
\code{\link{read_multsheets}()},
\code{\link{read_sheets}()}
}
\concept{readin functions}
