% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ods_multsheets.R
\name{read_ods_multsheets}
\alias{read_ods_multsheets}
\title{A wrapper around read_ods to read multiple sheets in a .ods file}
\usage{
read_ods_multsheets(filename)
}
\arguments{
\item{filename}{File to import}
}
\value{
A list of data frames--one data frame for each sheet
}
\description{
A wrapper around read_ods to read multiple sheets in a .ods file
}
\examples{
dat <- c("iris", "mtcars")
path <- "/home/em/Desktop/"

readODS::write_ods(get(dat[1]), paste(path, "x.ods"), row_names = FALSE, sheet = dat[1])

for(i in 2:length(dat)) {
readODS::write_ods(get(dat[i]), paste(path, "x.ods"), row_names = FALSE, sheet = dat[i],
  append = TRUE)}

read_ods_multsheets(paste(path, "x.ods"))

}
